+++ aeon/base/__init__.py
@@ -4,10 +4,10 @@ __all__ = [
     "BaseAeonEstimator",
     "BaseCollectionEstimator",
     "BaseSeriesEstimator",
-    "_ComposableEstimatorMixin",
+    "ComposableEstimatorMixin",
 ]
 
 from aeon.base._base import BaseAeonEstimator
 from aeon.base._base_collection import BaseCollectionEstimator
 from aeon.base._base_series import BaseSeriesEstimator
-from aeon.base._meta import _ComposableEstimatorMixin
+from aeon.base._meta import ComposableEstimatorMixin
