+++ aeon/base/tests/test_compose.py
@@ -1,0 +1,1 @@
+"""Test composable estimator mixin."""
