+++ aeon/classification/interval_based/_interval_forest.py
@@ -5,7 +5,7 @@ __all__ = ["IntervalForestClassifier"]
 
 import numpy as np
 
-from aeon.base.estimator.interval_based.base_interval_forest import BaseIntervalForest
+from aeon.base.estimators.interval_based.base_interval_forest import BaseIntervalForest
 from aeon.classification.base import BaseClassifier
