+++ aeon/classification/interval_based/_rise.py
@@ -5,7 +5,7 @@ __all__ = ["RandomIntervalSpectralEnsembleClassifier"]
 
 import numpy as np
 
-from aeon.base.estimator.interval_based.base_interval_forest import BaseIntervalForest
+from aeon.base.estimators.interval_based.base_interval_forest import BaseIntervalForest
 from aeon.classification import BaseClassifier
 from aeon.classification.sklearn import ContinuousIntervalTree
 from aeon.transformations.collection import (
