+++ aeon/regression/interval_based/_rise.py
@@ -5,7 +5,7 @@ __all__ = ["RandomIntervalSpectralEnsembleRegressor"]
 
 import numpy as np
 
-from aeon.base.estimator.interval_based.base_interval_forest import BaseIntervalForest
+from aeon.base.estimators.interval_based.base_interval_forest import BaseIntervalForest
 from aeon.regression import BaseRegressor
 from aeon.transformations.collection import (
     AutocorrelationFunctionTransformer,
