+++ docs/api_reference/base.md
@@ -1,0 +1,18 @@
+# Base
+
+The :mod:`aeon.base` module contains abstract base classes.
+
+## Base classes
+
+```{currentmodule} aeon.base
+```
+
+```{autosummary}
+:toctree: auto_generated/
+:template: class.rst
+
+BaseAeonEstimator
+BaseCollectionEstimator
+BaseSeriesEstimator
+Composable
+```
