--- docs/api_reference/base.rst
@@ -1,35 +1,0 @@
-.. _base_ref:
-
-Base
-====
-
-The :mod:`aeon.base` module contains abstract base classes.
-
-.. automodule:: aeon.base
-    :no-members:
-    :no-inherited-members:
-
-Base classes
-------------
-
-.. currentmodule:: aeon.base
-
-.. autosummary::
-    :toctree: auto_generated/
-    :template: class.rst
-
-    BaseAeonEstimator
-    BaseCollectionEstimator
-    BaseSeriesEstimator
-
-Estimator base classes
-----------------------
-
-.. currentmodule:: aeon.base.estimator
-
-.. autosummary::
-    :toctree: auto_generated/
-    :template: class.rst
-
-    hybrid.BaseRIST
-    interval_based.BaseIntervalForest
